Indent with tabs
Indent should be independent on user's preferences and tooling setup.
It's been long time since tab was restricted to 8 spaces only.

Editors allow to customize it to user's preferences so source code appearance
adapts to reader's preferences as well as code layout will not change when
reviewing result of tools which do not support such customization.